install on ruby 2.4
Basically the `mysql` gem is deprecated and no longer supported. In order to support 2.4.1 we need to remove this and fix anything that breaks when only having `mysql2` installed.

Converted the 2 scripts still relying on the `mysql` gem to use `mysql2`.

Simple tests locally using a single node cluster in docker:
```
$ ./bin/check-percona-cluster-size.rb -h 172.17.0.3 -p eb5840e9883418c0 -e 1
CheckPerconaClusterSize OK: Expected to find 1 nodes and found those 1
```

```
$ ./bin/check-wsrep-ready.rb -h 172.17.0.3 -p eb5840e9883418c0
CheckWsrepReady OK: Cluster is OK!
```

Signed-off-by: Ben Abrams <me@benabrams.it>